Bugfix report generation of incomplete bam annotations

+++ core/makeSummary.sh
@@ -127,19 +127,25 @@ function bamAnnotate {
 	BAMANNOUT=runStats/bamann/$(echo ${DIR[@]}|sed 's/ /_/g')_${2}.ggplot
 	BAMANNIMAGE=${BAMANNOUT/ggplot/pdf}
 	if [ ! -f $BAMANNOUT ]; then mkdir -p $( dirname $BAMANNOUT); fi
-	cat ${1}/*.anno.stats | head -n 1 | gawk '{print "type "$0" sample"}'  >$BAMANNOUT
-	for i in $(ls ${1}/*.anno.stats); do
-		name=$(basename $i)
-		arrIN=(${name//.$ASD/ })
-		grep sum $i | gawk -v x=${arrIN[0]} '{print $0" "x}';
-	done >> $BAMANNOUT
-	sed -i -r 's/\s+/ /g' $BAMANNOUT
-	Rscript ${NGSANE_BASE}/tools/bamann.R $BAMANNOUT $BAMANNIMAGE "Genome Features ${2}"
-	convert $BAMANNIMAGE ${BAMANNIMAGE/pdf/jpg}
-	echo "<h3>Annotation of mapped reads</h3>" >> $3
-	echo "<div><a href=$PROJECT_RELPATH/$BAMANNIMAGE><img src=\""$PROJECT_RELPATH/${BAMANNIMAGE/.pdf/}"-0.jpg\" width='250px' style='float:left;'><img src=\""$PROJECT_RELPATH/${BAMANNIMAGE/.pdf/}"-1.jpg\" width='250px' style='float:left;'></a></div>">>$3
-
-#	    python ${NGSANE_BASE}/tools/makeBamHistogram.py "${1}" $ROUTH >>$3
+	
+	for PROJECT in ${1}; do 
+    	if [ $(ls ${PROJECT}/*.anno.stats 2> /dev/null | wc -l ) -gt 0 ]; then 
+            echo "annotate ${PROJECT}"
+            cat ${PROJECT}/*.anno.stats | head -n 1 | gawk '{print "type "$0" sample"}' > $BAMANNOUT
+            for i in $(ls ${PROJECT}/*.anno.stats); do
+                name=$(basename $i)
+                arrIN=(${name//.$ASD/ })
+                grep --no-messages sum $i | gawk -v x=${arrIN[0]} '{print $0" "x}';
+        	done >> $BAMANNOUT
+        	sed -i -r 's/\s+/ /g' $BAMANNOUT
+        	Rscript ${NGSANE_BASE}/tools/bamann.R $BAMANNOUT $BAMANNIMAGE "Genome Features ${2}"
+        	convert $BAMANNIMAGE ${BAMANNIMAGE/pdf/jpg}
+        	echo "<h3>Annotation of mapped reads</h3>" >> $3
+        	echo "<div><a href=$PROJECT_RELPATH/$BAMANNIMAGE><img src=\""$PROJECT_RELPATH/${BAMANNIMAGE/.pdf/}"-0.jpg\" width='250px' style='float:left;'><img src=\""$PROJECT_RELPATH/${BAMANNIMAGE/.pdf/}"-1.jpg\" width='250px' style='float:left;'></a></div>">>$3
+    
+    #	    python ${NGSANE_BASE}/tools/makeBamHistogram.py "${PROJECT}" $ROUTH >>$3
+        fi
+    done
 
 }
 
@@ -354,7 +360,7 @@ if [[ -n "$RUNTOPHAT" || -n "$RUNTOPHATCUFFHTSEQ" ]]; then
     for dir in ${DIR[@]}; do
     	vali=$vali" $OUT/$dir/$TASKTOPHAT/"
     	cd $OUT/$dir/$TASKTOPHAT
-    	for d in $(find . -maxdepth 1 -mindepth 1 -type d -exec basename '{}' \; | grep "RNASeQC"); do
+    	for d in $(find . -maxdepth 1 -mindepth 1 -type d -exec basename '{}' \; | grep --no-messages "RNASeQC"); do
             echo "<a href=\"$PROJECT_RELPATH/$dir/$TASKTOPHAT/$d/index.html\">RNAseq-QC for $dir/$d</a><br/>" >> $CURDIR/$SUMMARYTMP
 		done
     done
